tox_requirements.txt
pip
cython==0.23.1
configargparse==0.9.3
numpy==1.9.2
pyyaml==3.11
pep8==1.7.0
flake8==2.4.1
pytest==2.7.3
pytest-cov==2.0.0
posix_ipc==1.0.0
pillow==2.9.0
pylint==1.4.4
sphinx==1.3.1
h5py==2.5.0
future==0.15.2
pycuda>=2015.1
scikit-cuda==0.5.1
pytools
